+++ pom.xml
@@ -123,7 +123,6 @@
                             <id>validate</id>
                             <phase>validate</phase>
                             <configuration>
-                            	<skip>true</skip>
                                 <configLocation>
                                     https://raw.githubusercontent.com/wso2/code-quality-tools/master/checkstyle/checkstyle.xml
                                 </configLocation>
