--- src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseEngine/Aggregator.java
@@ -1,8 +1,0 @@
-package io.datasearch.diseasedata.store.dengdipipeline.fuseEngine;
-
-/**
- * For aggregating data.
- */
-public class Aggregator {
-	public void aggregate() {};
-}
