--- src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseEngine/FuseEngine.java
@@ -1,13 +1,0 @@
-package io.datasearch.diseasedata.store.dengdipipeline.fuseEngine;
-
-/**
- * For data fusion.
- */
-public class FuseEngine {
-	//aggregating
-	private Aggregator aggregator;
-	//granularityConvertor
-	private GranularityConvertor granularityConvertor;
-	//Transformer
-	private Transformer transformer;
-}
