--- src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseEngine/GranularityConvertor.java
@@ -1,9 +1,0 @@
-package io.datasearch.diseasedata.store.dengdipipeline.fuseEngine;
-
-/**
- * For granularity conversion.
- */
-public class GranularityConvertor {
-	public void temporalConversion() {};
-	public void spatialConversion() {};
-}
