--- src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseEngine/Transformer.java
@@ -1,7 +1,0 @@
-package io.datasearch.diseasedata.store.dengdipipeline.fuseEngine;
-
-/**
- * to apply transformations.
- */
-public class Transformer {
-}
