+++ src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseengine/Aggregator.java
@@ -1,0 +1,8 @@
+package io.datasearch.diseasedata.store.dengdipipeline.fuseengine;
+
+/**
+ * For aggregating data.
+ */
+public class Aggregator {
+ public void aggregate() {}
+}
