+++ src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseengine/FuseEngine.java
@@ -1,0 +1,14 @@
+package io.datasearch.diseasedata.store.dengdipipeline.fuseengine;
+
+/**
+ * For data fusion.
+ */
+public class FuseEngine {
+ //aggregating
+ private Aggregator aggregator;
+ //granularityConvertor
+ private GranularityConvertor granularityConvertor;
+ //Transformer
+ private Transformer transformer;
+}
+
