+++ src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseengine/GranularityConvertor.java
@@ -1,0 +1,9 @@
+package io.datasearch.diseasedata.store.dengdipipeline.fuseengine;
+
+/**
+ * For granularity conversion.
+ */
+public class GranularityConvertor {
+ public void temporalConversion() { }
+ public void spatialConversion() { }
+}
