+++ src/main/java/io/datasearch/diseasedata/store/dengdipipeline/fuseengine/Transformer.java
@@ -1,0 +1,6 @@
+package io.datasearch.diseasedata.store.dengdipipeline.fuseengine;
+
+/**
+ * to apply transformations.
+ */
+public class Transformer { }
